Remove no-longer-relevant minimum PHP requirement
This module previously defined a minimum PHP version requirement of > 5.6. Since this branch of the module works with Drupal 8/9/10, and all of those have minimum PHP version of > 7, there is no value in specifying a minimum version lower than that. Drupal 10 requires a minimum PHP version of 8.1, but we cannot set that as the minimum PHP version for this module, since we also support Drupal 8/9. Conclusion: do not define a minimum PHP version.